adk-agents/adk_observability/requirements.txt
google-adk>=1.5.0 
openinference-instrumentation-google-adk>=0.1.3
arize-phoenix-otel>=0.13.0
feedparser>=6.0.11